Test for attribute responds to all attribute types.
Previously, the __isset() method returned true only for attributes and aliased
attributes, but would not return true for getter methods or associations,
which are tested when accessing data from the model directly in a template
language such as Twig. This was previously reported in
https://github.com/kla/php-activerecord/issues/156, but the fix was not
implemented correctly (__isset() should return a Boolean).